# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ai-test api",
    "version": "1.0.0",
    "description": "Hono API for the ai-test monorepo."
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RootResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ai-test api"
          },
          "status": {
            "type": "string",
            "example": "ok"
          }
        },
        "required": [
          "name",
          "status"
        ]
      },
      "HealthResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "uptime": {
            "type": "number",
            "example": 12.34
          }
        },
        "required": [
          "status",
          "uptime"
        ]
      },
      "HelloResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Hello, world!"
          }
        },
        "required": [
          "message"
        ]
      },
      "NoteList": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "user_id": {
                  "type": "string",
                  "format": "uuid"
                },
                "title": {
                  "type": "string"
                },
                "content": {
                  "type": "string"
                },
                "created_at": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "user_id",
                "title",
                "content",
                "created_at"
              ]
            }
          }
        },
        "required": [
          "notes"
        ]
      },
      "Note": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "user_id",
          "title",
          "content",
          "created_at"
        ]
      },
      "CreateNote": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "content": {
            "type": "string",
            "maxLength": 10000,
            "default": ""
          }
        },
        "required": [
          "title"
        ]
      }
    },
    "parameters": {}
  },
  "paths": {
    "/": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Service info",
        "responses": {
          "200": {
            "description": "Basic service info",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RootResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "tags": [
          "meta"
        ],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Health status and process uptime",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/hello": {
      "get": {
        "tags": [
          "greeting"
        ],
        "summary": "Greet by name",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "Railway"
            },
            "required": false,
            "name": "name",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "A greeting message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HelloResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/notes": {
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "List the authenticated user's notes",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "The notes belonging to the authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteList"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          }
        }
      },
      "post": {
        "tags": [
          "notes"
        ],
        "summary": "Create a note for the authenticated user",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNote"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created note",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Note"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token"
          }
        }
      }
    }
  },
  "webhooks": {}
}